+++ packages/playwright-msw/src/index.ts
@@ -1,3 +1,1 @@
 export * from "./fixture";
-export * from "./server";
-export * from "./types";
